+++ packages/create-twilio-function/src/create-twilio-function.js
@@ -2,7 +2,7 @@ const { promptForAccountDetails } = require('./create-twilio-function/prompt');
 const {
   createDirectory,
   createEnvFile,
-  createExampleFunction,
+  createExampleFromTemplates,
   createPackageJSON,
   createNvmrcFile
 } = require('./create-twilio-function/create-files');
@@ -14,31 +14,46 @@ const {
 const successMessage = require('./create-twilio-function/success-message');
 const ora = require('ora');
 const boxen = require('boxen');
+const { downloadTemplate } = require('twilio-run/dist/templating/actions');
+const { promisify } = require('util');
+const rimraf = promisify(require('rimraf'));
+
+async function cleanUpAndExit(projectDir, spinner, errorMessage) {
+  spinner.fail(errorMessage);
+  spinner.start('Cleaning up project directories and files');
+  await rimraf(projectDir);
+  spinner.stop().clear();
+  process.exitCode = 1;
+}
 
 async function createTwilioFunction(config) {
   const projectDir = `${config.path}/${config.name}`;
+  const spinner = ora();
 
   try {
+    spinner.start('Creating project directory');
     await createDirectory(config.path, config.name);
+    spinner.succeed();
   } catch (e) {
     switch (e.code) {
       case 'EEXIST':
-        console.error(
+        spinner.fail(
           `A directory called '${
             config.name
           }' already exists. Please create your function in a new directory.`
         );
         break;
       case 'EACCES':
-        console.error(
+        spinner.fail(
           `You do not have permission to create files or directories in the path '${
             config.path
           }'.`
         );
         break;
       default:
-        console.error(e.message);
+        spinner.fail(e.message);
     }
+    process.exitCode = 1;
     return;
   }
 
@@ -50,28 +65,58 @@ async function createTwilioFunction(config) {
   config = { ...accountDetails, ...config };
 
   // Scaffold project
-  const spinner = ora();
   spinner.start('Creating project directories and files');
-  await createDirectory(projectDir, 'functions');
-  await createDirectory(projectDir, 'assets');
+
   await createEnvFile(projectDir, {
     accountSid: config.accountSid,
     authToken: config.authToken
   });
   await createNvmrcFile(projectDir);
-  await createExampleFunction(`${projectDir}/functions`);
   await createPackageJSON(projectDir, config.name);
-  spinner.succeed();
+  if (config.template) {
+    spinner.succeed();
+    spinner.start(`Downloading template: "${config.template}"`);
+    await createDirectory(projectDir, 'functions');
+    await createDirectory(projectDir, 'assets');
+    try {
+      await downloadTemplate(config.template, '', projectDir);
+      spinner.succeed();
+    } catch (err) {
+      await cleanUpAndExit(
+        projectDir,
+        spinner,
+        `The template "${config.template}" doesn't exist`
+      );
+      return;
+    }
+  } else {
+    await createExampleFromTemplates(projectDir);
+    spinner.succeed();
+  }
 
   // Download .gitignore file from https://github.com/github/gitignore/
-  spinner.start('Downloading .gitignore file');
-  await createGitignore(projectDir);
-  spinner.succeed();
+  try {
+    spinner.start('Downloading .gitignore file');
+    await createGitignore(projectDir);
+    spinner.succeed();
+  } catch (err) {
+    cleanUpAndExit(projectDir, spinner, 'Could not download .gitignore file');
+    return;
+  }
 
   // Install dependencies with npm
-  spinner.start('Installing dependencies');
-  await installDependencies(projectDir);
-  spinner.succeed();
+  try {
+    spinner.start('Installing dependencies');
+    await installDependencies(projectDir);
+    spinner.succeed();
+  } catch (err) {
+    spinner.fail();
+    console.log(
+      `There was an error installing the dependencies, but your project is otherwise complete in ./${
+        config.name
+      }`
+    );
+  }
 
   // Success message