+++ packages/create-twilio-function/src/create-twilio-function/versions.js
@@ -1,4 +1,4 @@
 module.exports = {
-  twilioRun: '^1.0.1',
+  twilioRun: '2.* || >2.0.0-rc',
   node: '8.10.0'
 };